feat: congrats page performance
