+++ src/app/congrats/[resource_id]/components/index.ts
@@ -1,0 +1,2 @@
+export * from './congratsScreen'
+export * from './rrssButton'
