+++ src/app/congrats/[resource_id]/page.tsx
@@ -1,46 +1,36 @@
-'use client'
-
-import { Suspense, use, useEffect, useState } from 'react'
-import { CongratsScreen } from '@/components/custom/congratsScreen'
+import { CongratsScreen } from '@/app/congrats/[resource_id]/components'
 import { Resource } from '@/app/schema'
-import { SplashLoaderModal } from '@/components/custom/splashLoaderModal'
 
-export default function CongratsPage({ params }: { params: Promise<{ resource_id: string }> }) {
-    const { resource_id: resourceId } = use(params)
-    const [loading, setLoading] = useState<boolean>(true)
-    const [resource, setResource] = useState<Resource | null>(null)
+async function getResource(resourceId: string): Promise<Resource | null> {
+  try {
+    const response = await fetch(
+      `${process.env.APP_URL}/api/resources?resource_id=${resourceId}`,
+      { cache: 'no-store' } 
+    );
+    if (!response.ok) return null;
 
-  useEffect(() => {
-    if(resourceId) {
-      setLoading(true)
-      fetchResource()
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  const fetchResource = async () => {
-
-    try { 
-      const response = await fetch(`/api/resources?resource_id=${resourceId}`)
-      
-      if(!response.ok) {
-        throw new Error('No se pudo obtener el recurso')
-      }
-      const { data } = await response.json()
-
-    if(data) {
-      setResource(data[0])
-      setLoading(false)
-    }
+    const { data } = await response.json();
+    return data?.[0] || null;
   } catch (error) {
-    console.error('Error en /api/congrats:', (error as Error).message)
-    setLoading(false)
+    console.error("Error al obtener recurso:", error);
+    return null;
   }
-  }
+}
+
+export default async function CongratsPage({ params }: { params: Promise<{ resource_id: string }> }) {
+  const { resource_id } = await params
+  const resource = await getResource(resource_id)
+
+  if (!resource) {
+      return (
+        <div className="p-8 text-center text-red-500">
+          No se pudo encontrar el recurso solicitado.
+        </div>
+      )
+    }
+
 
   return (
-    <Suspense fallback={<SplashLoaderModal open={loading} message="Cargando..." />}>
       <CongratsScreen resourceName={resource?.name || ''} />
-    </Suspense>
   )
 }
